Fixed delete button in user management
